update comparison chart
